Refactor rendering pipeline: .Make sample distance multiplier a property of the viewport. move sample distance change out of updateRenderingPipeline method and add to VolumeViewport3D.
