+++ packages/core/src/RenderingEngine/VolumeViewport3D.ts
@@ -4,13 +4,13 @@ import { OrientationAxis, Events } from '../enums';
 import cache from '../cache/cache';
 import setDefaultVolumeVOI from './helpers/setDefaultVolumeVOI';
 import triggerEvent from '../utilities/triggerEvent';
-import { isImageActor } from '../utilities/actorCheck';
+import { actorIsA, isImageActor } from '../utilities/actorCheck';
 import { setTransferFunctionNodes } from '../utilities/transferFunctionUtils';
 import type vtkVolume from '@kitware/vtk.js/Rendering/Core/Volume';
 import type { ViewportInput } from '../types/IViewport';
 import type { ImageActor } from '../types/IActor';
 import BaseVolumeViewport from './BaseVolumeViewport';
-
+import type { Types } from '@cornerstonejs/core';
 /**
  * An object representing a 3-dimensional volume viewport. VolumeViewport3Ds are used to render
  * 3D volumes in their entirety, and not just load a single slice at a time.
@@ -34,6 +34,44 @@ class VolumeViewport3D extends BaseVolumeViewport {
       this.applyViewOrientation(orientation);
     }
   }
+
+  public setSampleDistanceMultiplier = (multiplier: number): void => {
+    console.debug(
+      `Setting sample distance multiplier in volumeviewport3d: ${multiplier}`
+    );
+    const actors = this.getActors();
+    actors.forEach((actorEntry) => {
+      if (actorIsA(actorEntry, 'vtkVolume')) {
+        const actor = actorEntry.actor as Types.VolumeActor;
+        const mapper = actor.getMapper();
+
+        if (mapper && mapper.getInputData) {
+          const imageData = mapper.getInputData();
+
+          if (imageData) {
+            const spacing = imageData.getSpacing();
+
+            //Calculate sample distance
+            const defaultSampleDistance =
+              (spacing[0] + spacing[1] + spacing[2]) / 6;
+
+            const sampleDistanceMultiplier = multiplier || 1;
+            let sampleDistance =
+              defaultSampleDistance * sampleDistanceMultiplier;
+
+            // Apply sample distance if specified
+            if (sampleDistance !== undefined && mapper.setSampleDistance) {
+              const currentSampleDistance = mapper.getSampleDistance();
+              mapper.setSampleDistance(sampleDistance);
+            }
+            console.debug(`New sample distance set: ${sampleDistance}`);
+          }
+        }
+      }
+    });
+
+    this.render();
+  };
 
   public getNumberOfSlices = (): number => {
     return 1;
